RestaurantsPage로 분리
기존의 컴포넌트들로 화면에 보여주는 레스토랑 목록 페이지를 분리
